# server/ee/libs/embedded/embedded-configuration/embedded-configuration-rest/embedded-configuration-rest-impl/openapi.yaml
...
openapi: "3.0.1"
info:
  title: "The Embedded Configuration Internal API"
  version: "1"
servers:
  - url: "/api/embedded/internal"
...
paths:
...
  /connected-user-project-workflows/{workflowReferenceCode}:
    get:
      description: "Get connected user project workflow."
      summary: "Get connected user project workflow."
      tags:
        - "connected-user-project-workflow"
      operationId: "getConnectedUserProjectWorkflow"
      parameters:
        - name: "X-Environment"
          description: "The environment."
          in: "header"
          required: false
          schema:
            $ref: '#/components/schemas/Environment'
        - name: "workflowReferenceCode"
          description: "The workflow reference code."
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "The connected user project workflow object."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectedUserProjectWorkflow"
  /connected-user-project-workflows/{workflowReferenceCode}/publish:
    post:
      description: "Publishes existing connected user project workflow."
      summary: "Publishes existing connected user project workflow"
      tags:
        - "connected-user-project-workflow"
      operationId: "publishConnectedUserProjectWorkflow"
      parameters:
        - name: "X-Environment"
          description: "The environment."
          in: "header"
          required: false
          schema:
            $ref: '#/components/schemas/Environment'
        - name: "workflowReferenceCode"
          description: "The workflow reference code."
          in: "path"
          required: true
          schema:
            type: "string"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  description: The description of a connected user project workflow version.
                  type: "string"
      responses:
        "204":
          description: "Successful operation."
  /connected-user-project-workflows/{workflowReferenceCode}/enable/{enable}:
    patch:
      description: "Enable/disable a connected user project workflow."
      summary: "Enable/disable a connected user project workflow"
      tags:
        - "connected-user-project-workflow"
      operationId: "enableConnectedUserProjectWorkflow"
      parameters:
        - name: "X-Environment"
          description: "The environment."
          in: "header"
          required: false
          schema:
            $ref: '#/components/schemas/Environment'
        - name: "workflowReferenceCode"
          description: "The workflow reference code."
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "enable"
          description: "Enable/disable the project deployment."
          in: "path"
          required: true
          schema:
            type: "boolean"
      responses:
        "204":
          description: "Successful operation."
...
components:
  schemas:
...
    ConnectedUserProjectWorkflow:
      type: "object"
      properties:
        connectedUserId:
          description: "The connected user id"
          type: "integer"
          format: "int64"
        projectId:
          description: "The project id"
          type: "integer"
          format: "int64"
        workflow:
          $ref: "../../../../../../libs/platform/platform-configuration/platform-configuration-rest/platform-configuration-rest-impl/openapi.yaml#/components/schemas/Workflow"
    Environment:
      $ref: "../../../../../../libs/platform/platform-rest/openapi/components/schemas/objects/environment.yaml"
...
    Workflow:
      description: "The blueprint that describe the execution of a job."
      x-implements: ['com.bytechef.platform.configuration.web.rest.model.WorkflowModelAware']
      allOf:
        - $ref: "../../../../../../libs/platform/platform-configuration/platform-configuration-rest/platform-configuration-rest-impl/openapi.yaml#/components/schemas/Workflow"
        - type: "object"
          properties:
            integrationWorkflowId:
              description: "The integration workflow id"
              type: "integer"
              format: "int64"
              readOnly: true
            workflowReferenceCode:
              description: "The workflow reference code"
              type: "string"
              readOnly: true
...